feat(web): display project boundaries from geo json (idas-131)
- cleaned up unit tests
